requirements.txt
altair==5.4.1
appnope==0.1.3
asttokens==2.4.0
attrs==24.2.0
audioread==3.0.1
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
comm==0.2.2
debugpy==1.8.0
decorator==5.1.0
exceptiongroup==1.2.2
executing==2.0.0
filelock==3.16.1
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.10
importlib_metadata==6.7.0
ipykernel==6.25.2
ipython==8.17.1
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.5.0
jupyter_core==5.7.2
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.6
mdurl==0.1.2
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
narwhals==1.14.0
nest_asyncio==1.5.7
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
openai-whisper==20240930
packaging==23.2
pandas==2.2.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==11.0.0
platformdirs==4.0.0
pooch==1.8.2
prompt_toolkit==3.0.40
protobuf==5.28.3
psutil==5.9.5
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.0.0
pycparser==2.22
pydeck==0.9.1
pydub==0.25.1
Pygments==2.16.1
python-dateutil==2.8.2
pytz==2024.2
pyzmq==25.1.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
Resemblyzer==0.1.4
rich==13.9.4
rpds-py==0.21.0
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
smmap==5.0.1
soundfile==0.12.1
soxr==0.5.0.post1
stack-data==0.6.2
streamlit==1.40.1
streamlit-audiorec==0.1.3
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.8.0
toml==0.10.2
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tornado==6.4.1
tqdm==4.67.0
traitlets==5.12.0
typing==3.7.4.3
typing_extensions==4.10.0
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.8
webrtcvad==2.0.10
xgboost==2.1.2
zipp==3.17.0